Orientation

Meshes.OrientationType — Type

OrientationType

The different types of orientation of a ring. Possible values are CW and CCW.

Meshes.orientation — Function

orientation(geom, [method])

Returns the orientation of the geometry geom as either counter-clockwise (CCW) or clockwise (CW).

Optionally, specify the orientation method.

See also WindingOrientation, TriangleOrientation.

Meshes.OrientationMethod — Type

OrientationMethod

A method for finding the orientation of rings and polygons.

Meshes.WindingOrientation — Type

WindingOrientation()

A method for finding the orientatino of rings and polygons based on the winding number.

References

Balbes, R. and Siegel, J. 1990. A robust method for calculating the simplicity and orientation of planar polygons

Meshes.TriangleOrientation — Type

TriangleOrientation()

A method for finding the orientation of rings and polygons based on signed triangular areas.

References

Held, M. 1998. FIST: Fast Industrial-Strength Triangulation of Polygons