Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.053027909709456145, 0.32961907544551083, 0.6535799781934757)
 Vec(0.958381601559826, 0.8701496185552141, 0.8246904130212244)
 Vec(0.09412954395956541, 0.7852700807683736, 0.6938233533328897)
 Vec(0.21509959428291814, 0.577121800509175, 0.40035044635812733)
 Vec(0.13500067799171733, 0.29554303695092843, 0.4325502576336595)
 Vec(0.6677151436920135, 0.6031164373303047, 0.7120715157968078)
 Vec(0.5507758056231401, 0.6291686084638701, 0.5368047371819611)
 Vec(0.1590494136512176, 0.189822776761013, 0.938574885925533)
 Vec(0.8016716847084838, 0.020947527737745886, 0.1280835558476755)
 Vec(0.3123376951858232, 0.7366586155501343, 0.9113516788420876)
 ⋮
 Vec(0.9586300915484557, 0.7381488101694371, 0.12357112434354167)
 Vec(0.23090219828802616, 0.17284047081818865, 0.048639216805548946)
 Vec(0.4843300091927073, 0.5785838631597312, 0.6022843900251402)
 Vec(0.11182277121219897, 0.20989151783289972, 0.6424332546567723)
 Vec(0.7544011084711739, 0.2471656544937475, 0.8709456812614567)
 Vec(0.9103305772942449, 0.346820021737583, 0.27477366962479455)
 Vec(0.3131551600777732, 0.2611889204429121, 0.28862179857997006)
 Vec(0.508150861752397, 0.4147779853000131, 0.600281675193425)
 Vec(0.03801742846205136, 0.7209618490489116, 0.6517874373952406)