Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.6674301857072155, 0.45375869057348794, 0.5985265900509071)
 Vec(0.5896992970128798, 0.4495490527909113, 0.9332159899388124)
 Vec(0.0943091556992689, 0.30926487057478047, 0.9690886961505368)
 Vec(0.8122898472341852, 0.7281213592058636, 0.6528525157424651)
 Vec(0.14407358382515456, 0.45585015521761196, 0.346932128406248)
 Vec(0.37450058061719715, 0.6390043606029279, 0.2419910827925098)
 Vec(0.9943612562300215, 0.8691163001777397, 0.9068955574909835)
 Vec(0.6663253804159083, 0.5475622896859036, 0.8288576766462715)
 Vec(0.7587408336908743, 0.8757023178700724, 0.5906319821477992)
 Vec(0.17344570110392954, 0.9097374465996784, 0.7897824496240984)
 ⋮
 Vec(0.1755372573935008, 0.4955114599942968, 0.02331845073664096)
 Vec(0.5049965066219799, 0.9081904422013147, 0.2677820557606221)
 Vec(0.3877736977082017, 0.5878236727956302, 0.4182782846822133)
 Vec(0.5941958242792464, 0.5858015355287035, 0.06204907948411409)
 Vec(0.758002579245406, 0.7119711977272107, 0.9404657936416712)
 Vec(0.7721188138953005, 0.5192960458284284, 0.02397134159074088)
 Vec(0.07521309263145837, 0.30028332706829675, 0.031040339586861698)
 Vec(0.8317678831208278, 0.08561207761853284, 0.4174860230476076)
 Vec(0.6192774109142681, 0.7389840842561467, 0.05916442647561704)