Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.3154031740557769, 0.04150502931979827, 0.03751654954452133)
 Vec(0.04325119223387108, 0.9731601712789193, 0.6370062629727216)
 Vec(0.8935986213213363, 0.24674329542043782, 0.8840128613708934)
 Vec(0.7246843734361893, 0.6702473708297751, 0.5469825201706698)
 Vec(0.8939712849153473, 0.023838312035695752, 0.764345153615124)
 Vec(0.9444360407328066, 0.9876660199049118, 0.41573004426818305)
 Vec(0.4158511588826229, 0.7999489433551099, 0.30721369121442466)
 Vec(0.17616707199956827, 0.9205711681614575, 0.19173401536296142)
 Vec(0.26964357006231066, 0.7121512625530466, 0.4108056997018079)
 Vec(0.19901489762970714, 0.8062969305965881, 0.11697490440780622)
 ⋮
 Vec(0.008906672564962803, 0.220815807728439, 0.20174744962603397)
 Vec(0.43762887887553303, 0.4912522801626117, 0.7302891185058237)
 Vec(0.843885004606805, 0.7875300631474809, 0.8322739833658808)
 Vec(0.5010944858847477, 0.8953595578839894, 0.3029985771339957)
 Vec(0.9083756513796232, 0.5888576774084557, 0.3692851159015752)
 Vec(0.08470247304303125, 0.3894424711295523, 0.7641406129304117)
 Vec(0.9202215598783018, 0.037599119360519184, 0.6749599996606703)
 Vec(0.4176835511044067, 0.801505872941394, 0.7233622860438946)
 Vec(0.3157755800487735, 0.9600925025254156, 0.45414867134944903)