Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.3820631315243126, 0.024889909232039487, 0.1268276502713328)
 Vec(0.2409678665956302, 0.5264448613919117, 0.7579305251884481)
 Vec(0.46815057916623826, 0.7831117200759876, 0.4928434003423994)
 Vec(0.5397188529551515, 0.8655152095017795, 0.20130732184111133)
 Vec(0.932368574408876, 0.44308859389682864, 0.9734871155603685)
 Vec(0.06401822818013592, 0.1942930715200052, 0.7277801602821432)
 Vec(0.8255054960239461, 0.3472138736643495, 0.8957281487653456)
 Vec(0.6929246915307138, 0.9817734706416961, 0.3446433872433716)
 Vec(0.32742785064038005, 0.7671931745048729, 0.6279656259536825)
 Vec(0.17858660517197034, 0.997333650708311, 0.7077890212880893)
 ⋮
 Vec(0.421739187697011, 0.42093075537058255, 0.13121210351853474)
 Vec(0.4062790745220348, 0.012690680761150763, 0.5660136413125725)
 Vec(0.8800261764086914, 0.3145610230344521, 0.017171567917790842)
 Vec(0.587737744145927, 0.13674322467530176, 0.3076614166814986)
 Vec(0.6895406820626897, 0.9134990640259736, 0.5612964398992036)
 Vec(0.9424126544101409, 0.49571070983135357, 0.12311596471316677)
 Vec(0.21483732619693974, 0.5954001641594501, 0.05599595257514611)
 Vec(0.6128110678193756, 0.06230033373598154, 0.696805123172798)
 Vec(0.2411596510912032, 0.8535406583736709, 0.636458580671553)