Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.12619173879198653, 0.8501333487917822, 0.0816476488630351)
 Vec(0.021482078849618835, 0.9502350455763455, 0.7809476927707409)
 Vec(0.685722897298161, 0.5579240924395178, 0.2995586293111515)
 Vec(0.2995349504185628, 0.4754268951112587, 0.8306152379364302)
 Vec(0.5527788032001762, 0.47715857930595273, 0.12413834975456983)
 Vec(0.48799054934894737, 0.9544765144317233, 0.8746563990884199)
 Vec(0.16973694152684293, 0.41313089101125566, 0.5963596149882677)
 Vec(0.8688752897227189, 0.40849841521772623, 0.29763844097005476)
 Vec(0.24765243182320018, 0.07954215863173075, 0.19482310958491733)
 Vec(0.6020944056578142, 0.011765709579620531, 0.24936609120330078)
 ⋮
 Vec(0.4021035962781223, 0.30883783878543714, 0.6627200600984385)
 Vec(0.2983494845153345, 0.6491892242447869, 0.2309491997167944)
 Vec(0.9546289336093678, 0.6873537932273744, 0.008493969381423216)
 Vec(0.30548529047281103, 0.5950893261951736, 0.31344501914470735)
 Vec(0.03252474367280056, 0.8446415072769908, 0.10266234413480646)
 Vec(0.19009108899352334, 0.36462526639812154, 0.32408339847080725)
 Vec(0.9625791893201904, 0.11972208569940745, 0.7069346007211663)
 Vec(0.6202750518532015, 0.14539621243636836, 0.850674143227554)
 Vec(0.8870399047997058, 0.5308843084269482, 0.9815356084571822)