Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.9822493426293998, 0.9993884061823473, 0.9512986504942299)
 Vec(0.03269532649804163, 0.10088715083446964, 0.698819568389641)
 Vec(0.019715704052560512, 0.037465218413495815, 0.8606806761873119)
 Vec(0.40237015929521947, 0.019026766523287164, 0.9874351507769282)
 Vec(0.1699555546166327, 0.5714489704277173, 0.8116508953950686)
 Vec(0.6316680422489416, 0.2547976175109333, 0.2238698693625395)
 Vec(0.058492405078758525, 0.9064191336952361, 0.18062613542475614)
 Vec(0.5048853786101166, 0.8640013109938919, 0.23032708569619753)
 Vec(0.024476599839075908, 0.3550512728269801, 0.34270308336695643)
 Vec(0.30894663989032367, 0.9169846223502589, 0.4132699784710595)
 ⋮
 Vec(0.7643054490467895, 0.17655715140787986, 0.9702217484882799)
 Vec(0.2903148898399479, 0.08785213979258999, 0.0651107351299034)
 Vec(0.6787819466213922, 0.1850469291111546, 0.7784567445277505)
 Vec(0.39848858464571346, 0.07808987718656435, 0.8703178498531807)
 Vec(0.6220377026749327, 0.01646577453404008, 0.05084859477762638)
 Vec(0.253959526045178, 0.18094499245223528, 0.39091892736387324)
 Vec(0.7421901481250275, 0.6121259510852078, 0.1740918502929808)
 Vec(0.5450152697078269, 0.28874358323043103, 0.5796843899610143)
 Vec(0.7966474344731932, 0.08933832749879633, 0.1134756397251927)