Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7433308070173322, 0.8580181216717644, 0.07428265531258826)
 Vec(0.6372997976981224, 0.6241585512154979, 0.5417802268439453)
 Vec(0.8855598387318147, 0.747769036270881, 0.4403836163007432)
 Vec(0.6104890079868133, 0.8570364027242875, 0.37250465302990743)
 Vec(0.5197012095471307, 0.7625939705447059, 0.9084873480558968)
 Vec(0.9331511586327136, 0.04602601983621413, 0.2970435358538539)
 Vec(0.09119824554111722, 0.1469248496229657, 0.30875851297846546)
 Vec(0.6725237432460593, 0.1857744057221864, 0.9082405476808894)
 Vec(0.8815725078957752, 0.2284975705259532, 0.27934355055440996)
 Vec(0.3927073927419197, 0.7879878314825634, 0.4440798365307348)
 ⋮
 Vec(0.004317268763323057, 0.08940309958710246, 0.15217885106473095)
 Vec(0.21537053094762015, 0.4417767249809412, 0.9543433835818145)
 Vec(0.6687004662516759, 0.5529239588831941, 0.2450217233851799)
 Vec(0.6496550814591642, 0.9396376641524751, 0.4969437121253597)
 Vec(0.3186905976812402, 0.4320475699341262, 0.3661815930951303)
 Vec(0.12878145258871043, 0.8364621237191219, 0.6151003699583389)
 Vec(0.12124354822275119, 0.552992771699421, 0.9869221024836023)
 Vec(0.9925557422877712, 0.8473594442888466, 0.7109201911748512)
 Vec(0.9419133769089977, 0.1690931209115516, 0.6931627004713162)