Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.32323638630457074, 0.4039776017924773, 0.14452451884019935)
 Vec(0.6925822431159603, 0.14287278809810577, 0.12253238173463199)
 Vec(0.5257492586447442, 0.7600255857827477, 0.47388566250113007)
 Vec(0.8643528197756012, 0.42332003866869183, 0.01040884152791588)
 Vec(0.5441980374331187, 0.7505902681634488, 0.7398836817511737)
 Vec(0.2997817025896906, 0.8452191438882974, 0.5291578858292318)
 Vec(0.8886915334689945, 0.08943331036772084, 0.6706872404250517)
 Vec(0.06841998850428554, 0.019597772709986305, 0.07707343956381363)
 Vec(0.8031366977360657, 0.15904991561375426, 0.2789717594056089)
 Vec(0.9571197027047282, 0.21909263922840383, 0.3562129476072877)
 ⋮
 Vec(0.8558649887474055, 0.7799939563026529, 0.6074286775059492)
 Vec(0.6950691473107665, 0.09050675108230266, 0.29985383204195404)
 Vec(0.8286987982430932, 0.4531643390326727, 0.6813433642178292)
 Vec(0.7182579831248366, 0.17627206079489, 0.3081155516292201)
 Vec(0.059021456489864454, 0.25594644944641354, 0.7962423540488126)
 Vec(0.9245792761676083, 0.4896630656222871, 0.2964276218888049)
 Vec(0.431303981568084, 0.8195384334755467, 0.4384546665244208)
 Vec(0.892612024969454, 0.6052736366898984, 0.41869390102122683)
 Vec(0.3838761766908352, 0.06957421768192729, 0.7060297487284081)