Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.23314793952039803, 0.7010047803906814, 0.62191309369826)
 Vec(0.33483122432953405, 0.3858758974933545, 0.23969163642544244)
 Vec(0.11528544103661176, 0.7865252658597687, 0.7030299462470744)
 Vec(0.7315163204718714, 0.05577346637331104, 0.39836826473889253)
 Vec(0.3230112001257198, 0.8477366195224884, 0.7999856182435343)
 Vec(0.004354787730935117, 0.20016442546928948, 0.7891957946690291)
 Vec(0.4771296465002448, 0.04926628399818689, 0.30434306758146024)
 Vec(0.5136612361215456, 0.9045806489141112, 0.8774683796283965)
 Vec(0.11727315966989804, 0.6777708812031815, 0.7398060074210377)
 Vec(0.43077564967842275, 0.85403397749748, 0.7928064259820421)
 ⋮
 Vec(0.32249601437515285, 0.11606172816297067, 0.3127562239531728)
 Vec(0.9179821266333902, 0.3049591921910415, 0.7117221504268032)
 Vec(0.702680247135953, 0.3865890465688516, 0.3536444453188379)
 Vec(0.14437831884652153, 0.8665111774209485, 0.12093360303489287)
 Vec(0.39170957600557665, 0.9467025458013792, 0.9026906832159645)
 Vec(0.04098847061025779, 0.3640630337222177, 0.851584113834447)
 Vec(0.44471213868463844, 0.9080547798270469, 0.12740103971762717)
 Vec(0.7217118040101429, 0.3516369090227234, 0.2689059155621273)
 Vec(0.7582508027171869, 0.5647066686075847, 0.07435024060111162)