Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7725559272217998, 0.947625664075879, 0.41402266992873726)
 Vec(0.6707355842555504, 0.39976810853112354, 0.7669753737200334)
 Vec(0.11501168130575978, 0.9621865077396565, 0.7722614248845207)
 Vec(0.376835117308571, 0.44888671160754434, 0.7850861790604096)
 Vec(0.48982333689990143, 0.27495799809829924, 0.2500238057140983)
 Vec(0.7420188164623075, 0.556965502177531, 0.35686471070384806)
 Vec(0.10983409052119453, 0.2875333059316595, 0.7315632608737053)
 Vec(0.12720836686016024, 0.870852835900048, 0.37615590373698515)
 Vec(0.06901731666885325, 0.211357572741116, 0.19250489815446414)
 Vec(0.6535952743630518, 0.8265758890978384, 0.8554076554233548)
 ⋮
 Vec(0.18332338843306117, 0.8210785670319307, 0.1493103174577627)
 Vec(0.6380446397482923, 0.7482291654097395, 0.13727295299190545)
 Vec(0.1036078655280932, 0.8650100845548153, 0.4018642586738398)
 Vec(0.32283345957999743, 0.09822702304722031, 0.5959295823819979)
 Vec(0.38902166706017327, 0.9829422115764174, 0.3678671364628684)
 Vec(0.5978364197369355, 0.47758904854277906, 0.4265092142347918)
 Vec(0.0014754850650008366, 0.48999167686581446, 0.42704784172538945)
 Vec(0.9753134852624067, 0.6339516423833459, 0.9507241517089309)
 Vec(0.5988159792017275, 0.9539118341622724, 0.7133453449382136)