Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.5270891034109662, 0.8015826913964262, 0.019222638053313146)
 Vec(0.4357684021894197, 0.39472968866794267, 0.3694667583438179)
 Vec(0.423362559819644, 0.8595398475866248, 0.8433986624819355)
 Vec(0.10842645308936627, 0.18917343883694493, 0.35877967147181067)
 Vec(0.8183586202788963, 0.9143164398309538, 0.019437746566001723)
 Vec(0.9581224546002339, 0.6400994018543688, 0.6250579700528192)
 Vec(0.9588129137468993, 0.42081147116095596, 0.6583359901291992)
 Vec(0.6434908834057991, 0.18388250257143002, 0.5510648574667918)
 Vec(0.34584408957571067, 0.18042450954553024, 0.6726159651352007)
 Vec(0.3180244063524874, 0.1857782041510616, 0.9122935608947371)
 ⋮
 Vec(0.23468591715049414, 0.4604696532485736, 0.7707719602641845)
 Vec(0.6227179352187592, 0.4093049035906686, 0.5673328577267958)
 Vec(0.7932669076475575, 0.026851190499768696, 0.7248904211616891)
 Vec(0.1909353942943941, 0.3934101224330988, 0.8041247467987037)
 Vec(0.04207578287103919, 0.9334313008071626, 0.4945097027542117)
 Vec(0.7517506926635065, 0.2373455328610905, 0.27306941494410586)
 Vec(0.7493808580175807, 0.14021879248008762, 0.45500812743520613)
 Vec(0.9282748958189201, 0.6448063470946762, 0.6164714123346097)
 Vec(0.5712128772047866, 0.7840473573303876, 0.8478438608331583)