Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.26950280311177044, 0.24151104478327412, 0.42653714420412336)
 Vec(0.8002652800095063, 0.6786151692413922, 0.05117691316610373)
 Vec(0.9161621134812258, 0.0895531400768147, 0.7959945686233326)
 Vec(0.9324969320348746, 0.31412792840374637, 0.5578847456364587)
 Vec(0.9250447143483248, 0.2653581297316261, 0.8050575331593218)
 Vec(0.7163452876934552, 0.20009202378163482, 0.7651744839040305)
 Vec(0.4241158394350355, 0.883065066930786, 0.4465183076453647)
 Vec(0.7490544059811456, 0.19889998003184362, 0.6545840045019546)
 Vec(0.9161121258425958, 0.046581217131148045, 0.5432834991391284)
 Vec(0.5178941845589985, 0.06696047768749325, 0.8252599292622577)
 ⋮
 Vec(0.06221431846012693, 0.7775487220209886, 0.7169131616997988)
 Vec(0.11384801311647241, 0.6869947690402488, 0.428561030501462)
 Vec(0.9699597226708404, 0.2778314092282811, 0.12442501086248448)
 Vec(0.10236216959216848, 0.5713329347844113, 0.15262746315811337)
 Vec(0.8137690782816659, 0.8752563585515506, 0.31715814840420753)
 Vec(0.45312975324962046, 0.11853558548632825, 0.9676510742394306)
 Vec(0.6876031839788234, 0.5392681730937131, 0.6087781000035999)
 Vec(0.2819817468529783, 0.1786813914174057, 0.8085228112101619)
 Vec(0.18743991057502685, 0.32649510504406276, 0.7037583669643198)