Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.9944867708070321, 0.30610409969484054, 0.7396121267700732)
 Vec(0.48518305755899405, 0.48143062803515957, 0.9630461063858657)
 Vec(0.843227785152202, 0.337258673592993, 0.8554843704517842)
 Vec(0.8537223378204244, 0.8135431622863168, 0.7297819451828296)
 Vec(0.7407997142914453, 0.7999073147054349, 0.4341644968823236)
 Vec(0.835076680120419, 0.5828450346230405, 0.329141380756826)
 Vec(0.0966287590429602, 0.4399718948714729, 0.3543078723790771)
 Vec(0.04385785016498178, 0.1259548545672673, 0.7423691661422562)
 Vec(0.8916910592339068, 0.6506662991831565, 0.12313201860934575)
 Vec(0.6430208796186049, 0.23231490854888126, 0.36455973138142805)
 ⋮
 Vec(0.08417506458712898, 0.3705920102536542, 0.45077658722780944)
 Vec(0.24217436529868153, 0.528563279104269, 0.22407727496756125)
 Vec(0.5337562608549025, 0.4104570724149442, 0.2824109203159868)
 Vec(0.720461887299895, 0.6370301008975854, 0.7306300411087219)
 Vec(0.909452431911259, 0.05525921335571149, 0.22960861514815123)
 Vec(0.18039289727176133, 0.833347283394216, 0.3452310906205055)
 Vec(0.2837497850231452, 0.6350646056116075, 0.9756605392289585)
 Vec(0.1738643277028783, 0.4389204100770061, 0.8149343653663753)
 Vec(0.4887697720300881, 0.8098099060116939, 0.876985082283964)