Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.19598370678769783, 0.44947644249424334, 0.24837439546757556)
 Vec(0.6127525453873961, 0.45332403769302343, 0.049208149112288324)
 Vec(0.6886805488718055, 0.6115961765058413, 0.7891778534052624)
 Vec(0.21746182575102924, 0.7858362170184416, 0.2920619421521763)
 Vec(0.04267945612264645, 0.4500267261861943, 0.9577450870200546)
 Vec(0.9473048310336732, 0.7005807128528243, 0.2019174158775332)
 Vec(0.5383448429557848, 0.5690358818000668, 0.24470477856511352)
 Vec(0.20600387857815594, 0.3417518444268426, 0.17615033802067914)
 Vec(0.5706637490542918, 0.17616628607104823, 0.13832275497157842)
 Vec(0.3770473537661847, 0.5743255399122302, 0.8814784686590714)
 ⋮
 Vec(0.3803735174059659, 0.8295375244220486, 0.08540410189026426)
 Vec(0.2752922144461698, 0.8547165658322712, 0.17482025528722134)
 Vec(0.06251739393078792, 0.07884387691170991, 0.44697556247936)
 Vec(0.9226790029295243, 0.5334265349309893, 0.14227178995484446)
 Vec(0.20473647492371938, 0.8062629352469819, 0.7631288231557095)
 Vec(0.9204478247092699, 0.823046921516269, 0.5338278539839959)
 Vec(0.038455291126675895, 0.7917436649224945, 0.3877454858318946)
 Vec(0.5568450386543962, 0.3764561711584461, 0.09381962495881013)
 Vec(0.8413619792886445, 0.14207166286388273, 0.7382816543839709)