Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7003296154013079, 0.9610360710671574, 0.16502020357024139)
 Vec(0.4660581360102126, 0.6173673030196147, 0.21056201360386928)
 Vec(0.6824812735336916, 0.9842081387076942, 0.47637359062982465)
 Vec(0.14133620563773353, 0.5379450349705148, 0.8666110784163161)
 Vec(0.7436797369388749, 0.5229804785249538, 0.49683497481351035)
 Vec(0.9046149979433773, 0.7480259694029624, 0.9360598808922463)
 Vec(0.1248851233671393, 0.789869686652104, 0.984964040016397)
 Vec(0.32562929719494105, 0.599028453528949, 0.8158528401847253)
 Vec(0.3851510351250209, 0.8843340994718967, 0.7701997442050286)
 Vec(0.04167117671842313, 0.5639706764826232, 0.08796592330782482)
 ⋮
 Vec(0.44495158229762155, 0.3589643180856761, 0.0567753145323987)
 Vec(0.10792434126922779, 0.3410211927777329, 0.1995910542994216)
 Vec(0.21943273834272004, 0.4839132307344014, 0.943758952891405)
 Vec(0.9050178928882441, 0.8813720492452161, 0.5169368734912424)
 Vec(0.245892172015876, 0.7516810185183224, 0.23200889881158004)
 Vec(0.20392320470768055, 0.3522591198007853, 0.44772812228094383)
 Vec(0.6682158425033398, 0.6086027104617912, 0.8869090028601305)
 Vec(0.29203566642834555, 0.9884026238525065, 0.9497552442280218)
 Vec(0.7043069625572181, 0.9245496875332105, 0.9077765706281251)