Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.2165255408577771, 0.2992738117699272, 0.5015639996682858)
 Vec(0.9921057956314535, 0.026466781925154614, 0.6670517530468708)
 Vec(0.32996178141243526, 0.9040580089864421, 0.7350857819788581)
 Vec(0.7867318198595433, 0.7641199105937573, 0.5569364606337789)
 Vec(0.4173520735866054, 0.9756781301403679, 0.6708616064353099)
 Vec(0.23541076358890423, 0.9184844042128028, 0.3768670135961041)
 Vec(0.350116922131106, 0.11371052199166354, 0.3199958108735028)
 Vec(0.23467333466530205, 0.22776650773436535, 0.32680321385712496)
 Vec(0.5777095345266591, 0.41961489475661773, 0.0998581231642992)
 Vec(0.5094546940944599, 0.19457268422069685, 0.3560082771890162)
 ⋮
 Vec(0.9038449340568105, 0.3125764737331094, 0.8162219859701991)
 Vec(0.4222435304273985, 0.8291334744722363, 0.9599789608665563)
 Vec(0.8049485258240808, 0.4196370480716767, 0.8999534319375766)
 Vec(0.4968680083354359, 0.601540261644028, 0.8592267942937577)
 Vec(0.3913637147985467, 0.0014017945990876246, 0.725394518261407)
 Vec(0.6869513139436254, 0.05384319372021584, 0.09984585103465438)
 Vec(0.6357641937932068, 0.5999735980836961, 0.5204274372620704)
 Vec(0.2563207911685801, 0.06690388458427532, 0.7752939369895114)
 Vec(0.7501525422889899, 0.12231810758697748, 0.28644018407704974)