Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.6871511456963814, 0.8317378335541386, 0.04505960886833715)
 Vec(0.37630357631556133, 0.7974647686218364, 0.33049441680597047)
 Vec(0.9029004879769341, 0.5828042140383203, 0.25030784251060256)
 Vec(0.1640707595534695, 0.8724040947160604, 0.3069951108592147)
 Vec(0.7712399568623757, 0.7501906769378287, 0.09640906148617767)
 Vec(0.8767612327911287, 0.0673411856660956, 0.8124923998377085)
 Vec(0.9649929915656796, 0.017145705733909944, 0.6925167207879577)
 Vec(0.3820054284283392, 0.9468739963921764, 0.15197636344610033)
 Vec(0.969476643188006, 0.8177136733718638, 0.5552795544167713)
 Vec(0.9025676468108254, 0.4246506748182356, 0.5647339681064893)
 ⋮
 Vec(0.1329337853128847, 0.45789917059926544, 0.6782521365384226)
 Vec(0.4453613632065905, 0.6671347193537516, 0.029269265682451806)
 Vec(0.7164462248741852, 0.285093354026543, 0.5830610724852343)
 Vec(0.4197605241760076, 0.622242643387142, 0.6874147553757043)
 Vec(0.9410290160041699, 0.8510869640464075, 0.5145495758051418)
 Vec(0.4413538830100142, 0.4366400036954562, 0.9959282197499234)
 Vec(0.8833272481757908, 0.6772738590885625, 0.5231366647737816)
 Vec(0.8667552502948465, 0.7138177000628696, 0.1671731688229382)
 Vec(0.4096176341121772, 0.10718260362277487, 0.030112806115648816)