Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.18306855093640084, 0.9576195437810218, 0.978667857051466)
 Vec(0.029990222810836964, 0.9288474882599295, 0.4894538206567691)
 Vec(0.7188395135833843, 0.6518357584489891, 0.04391450739855329)
 Vec(0.8792443497193977, 0.33228229014045874, 0.31255003375522783)
 Vec(0.8324022398180332, 0.7297738579897912, 0.10673397412163121)
 Vec(0.1862134496299751, 0.9270245796394171, 0.52924407897589)
 Vec(0.024486700847366483, 0.16903610173367156, 0.43816967500999593)
 Vec(0.7172920899480892, 0.952430759475548, 0.16660759855088902)
 Vec(0.11692955087070989, 0.3993494680862085, 0.7994094032333683)
 Vec(0.981436420325985, 0.2807040216653722, 0.2197953275722856)
 ⋮
 Vec(0.4551991729407876, 0.23481204593680716, 0.07180963891300007)
 Vec(0.7186036459948494, 0.6381575061584771, 0.41366808815933276)
 Vec(0.36335185728586916, 0.4701239856292505, 0.8864036022938606)
 Vec(0.8550321753573928, 0.2881940288557394, 0.09784770020246625)
 Vec(0.3745404019106443, 0.964503926850185, 0.18092119925832661)
 Vec(0.3064557258137166, 0.29151410446224935, 0.9449222847576553)
 Vec(0.4738731119930242, 0.842481118416656, 0.5211464394282448)
 Vec(0.27523232905766015, 0.7472953044446919, 0.05046227221973043)
 Vec(0.9839879308107726, 0.39586520964739114, 0.590478267364149)