Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.8210499399156727, 0.8504982606692015, 0.6018362379128346)
 Vec(0.040359320608769345, 0.05500598613859342, 0.5892687944824856)
 Vec(0.700063860073638, 0.20239214026215135, 0.802933669478516)
 Vec(0.15736007939030983, 0.44222729067387234, 0.5628687694696908)
 Vec(0.8397455119622144, 0.7467683407063223, 0.0983377316467301)
 Vec(0.3482991195032893, 0.8989576414606185, 0.013966057928376574)
 Vec(0.7120158806520165, 0.7981705074653598, 0.2268868173308023)
 Vec(0.625567430873744, 0.15132234412925316, 0.12847547941249582)
 Vec(0.3571695589263899, 0.06628907995941546, 0.9077949919688689)
 Vec(0.7403579195973076, 0.9679632863235204, 0.2229103688686701)
 ⋮
 Vec(0.5176975256649651, 0.07845584925596116, 0.2567916942244808)
 Vec(0.9815435559119341, 0.0465979589807024, 0.24706897266615047)
 Vec(0.4802736894826086, 0.6122617063582753, 0.6512535646429929)
 Vec(0.9016034534123708, 0.6870370635135025, 0.17717610058919708)
 Vec(0.7811262194763231, 0.6263263380928156, 0.33200915777801043)
 Vec(0.47116523947904176, 0.7598671642918836, 0.9917750130380426)
 Vec(0.7801284630123119, 0.9328458221588269, 0.03179001475367804)
 Vec(0.8691250318138811, 0.30534154649540834, 0.7078232247951353)
 Vec(0.6778590561957053, 0.790330408545077, 0.22913821055133998)