Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.42305165173424586, 0.006112215923417819, 0.21723275201783454)
 Vec(0.49311926279303175, 0.6960967691723162, 0.06842973478749625)
 Vec(0.10278624494270516, 0.718600253500021, 0.2634075656055267)
 Vec(0.692620663112127, 0.7581632683325586, 0.9780192804151202)
 Vec(0.8406694807263534, 0.8023700310584014, 0.17840569163767872)
 Vec(0.42590573439299706, 0.1166714363758008, 0.5562098218204327)
 Vec(0.2549195335725044, 0.011908902206274274, 0.1910679884490617)
 Vec(0.2594926783352738, 0.6594846458575632, 0.9099583565842227)
 Vec(0.06199832877960554, 0.6637473560702184, 0.2257619297427761)
 Vec(0.5394893358835282, 0.38336647746726793, 0.8635289743868669)
 ⋮
 Vec(0.9549460782718318, 0.8609081254978964, 0.03223665711391843)
 Vec(0.9624842896147199, 0.981704893258676, 0.12417307393379795)
 Vec(0.1527012194937022, 0.7349438516768713, 0.1276697161189304)
 Vec(0.7201234288692854, 0.5820028595375596, 0.9818113164025998)
 Vec(0.28232027547301586, 0.3225707828819644, 0.1346973834458901)
 Vec(0.4297391551473092, 0.9514438859367247, 0.8372992863025495)
 Vec(0.6331304946210373, 0.6125793789577759, 0.6816505691364269)
 Vec(0.6382946882368035, 0.42796732891169276, 0.9107691094729663)
 Vec(0.9351378334977363, 0.7005248875989227, 0.9432448854417983)