Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.4464825236245814, 0.4171938146208899, 0.537835446460229)
 Vec(0.34542959422023267, 0.26670551041312396, 0.9324868710251281)
 Vec(0.3163072220245724, 0.19774571798612317, 0.9701123579237981)
 Vec(0.4770886239545279, 0.9567242448062178, 0.2876303194541556)
 Vec(0.028007857175293083, 0.39197512246276933, 0.47863055930885656)
 Vec(0.8255320702661346, 0.3535793962932955, 0.5489689276123518)
 Vec(0.9504721445302398, 0.16764317749225144, 0.13991640808635408)
 Vec(0.13929121048213378, 0.7855622245459096, 0.577850965299523)
 Vec(0.7506226342698298, 0.5328741277997392, 0.9143098988724317)
 Vec(0.9152567823087157, 0.050190908427092173, 0.1388390329322433)
 ⋮
 Vec(0.06263089214636086, 0.4725963189034931, 0.20914130262685116)
 Vec(0.751848397810296, 0.6944713713811226, 0.38037736139627465)
 Vec(0.3722239041291667, 0.903577589149753, 0.10421965376693598)
 Vec(0.3302484958995724, 0.6023385506578144, 0.7158462751433693)
 Vec(0.8720010830550929, 0.6484194944986573, 0.8266310350869809)
 Vec(0.03042994836902746, 0.12117144410575043, 0.9627412951838206)
 Vec(0.07871870573815387, 0.19901742898308794, 0.9590637035938611)
 Vec(0.4465428152749902, 0.9774881679739966, 0.7221835673155991)
 Vec(0.5446085775707753, 0.21309218513085237, 0.3270583311775118)