Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.12546533672924642, 0.6934057597155984, 0.46184676705063765)
 Vec(0.13393926909453957, 0.7923937289321918, 0.2357773722792138)
 Vec(0.2761199402586859, 0.7031793132804748, 0.9444147755835913)
 Vec(0.41742737883729686, 0.9526765049967487, 0.8937403655570635)
 Vec(0.9411132212459451, 0.19899239958784298, 0.9356126441253404)
 Vec(0.9115248631649275, 0.5049745498999014, 0.26500967059533254)
 Vec(0.5622186862904929, 0.175300401140012, 0.8528430058521828)
 Vec(0.2204977753848112, 0.8746159909282211, 0.8066254822004347)
 Vec(0.01219162368145843, 0.6447233448718985, 0.1383501616563253)
 Vec(0.7403488932389173, 0.8400106628865285, 0.06401312925142755)
 ⋮
 Vec(0.6097825317682518, 0.11814513526231996, 0.6887850280249105)
 Vec(0.6526864791053139, 0.8052269292719765, 0.6237998212844982)
 Vec(0.9629876430489961, 0.8866481719156372, 0.14340325022896905)
 Vec(0.5140033356726987, 0.10236935789855495, 0.402458876504007)
 Vec(0.48451186849209915, 0.7548714226112945, 0.842138076019041)
 Vec(0.12725843481299015, 0.5201374479206826, 0.33240436578758636)
 Vec(0.6400819807057051, 0.8928865256934865, 0.2853270541695444)
 Vec(0.7408457092605321, 0.3203740321656853, 0.27535549129218295)
 Vec(0.27684398763353046, 0.8882745957920094, 0.648988237434173)