Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.48403126691877485, 0.23648459291173496, 0.615072028415779)
 Vec(0.38689229572393224, 0.6289252454249664, 0.24808384550089035)
 Vec(0.0754021755511064, 0.03247604159705075, 0.7507099028280165)
 Vec(0.2985474403818775, 0.018159815271337254, 0.4821811768700328)
 Vec(0.014822598522964303, 0.9918202522705856, 0.5557571709371459)
 Vec(0.8603760936050537, 0.38247773461631474, 0.573076062513795)
 Vec(0.2304259747034776, 0.5118037488371906, 0.37970897139276993)
 Vec(0.5694402027787295, 0.8588447059854851, 0.23465131630462777)
 Vec(0.5863486371777841, 0.37244936380536975, 0.6221061833134498)
 Vec(0.37880928347151277, 0.3957824154673315, 0.5838810761350388)
 ⋮
 Vec(0.1022677414181351, 0.7278370155699846, 0.9692862233032076)
 Vec(0.37287409071075384, 0.9806839877567495, 0.5599649055789494)
 Vec(0.4217167415911063, 0.7324522707813754, 0.033960429035524986)
 Vec(0.4503459895976115, 0.7480998415148136, 0.11765706187372649)
 Vec(0.1423462807973942, 0.22137350385014387, 0.7174486214489977)
 Vec(0.07634038867716497, 0.6176488286735663, 0.5417225021756377)
 Vec(0.6276949919575241, 0.062154957609995454, 0.8737673357696478)
 Vec(0.8330012632497422, 0.6346699986927532, 0.9102008263246933)
 Vec(0.342037569768719, 0.63767317292358, 0.38508066047002953)