Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7463833454839232, 0.42245281369661125, 0.6509104793008638)
 Vec(0.4516235535551778, 0.34592480246428425, 0.9571121522153412)
 Vec(0.26983855348498387, 0.18913746239034634, 0.35236885633222825)
 Vec(0.8384029355319521, 0.9975200645437831, 0.83096430269067)
 Vec(0.23459161017751728, 0.9929653601131343, 0.39393446923344067)
 Vec(0.006341730376172472, 0.424057997354636, 0.8996980033865856)
 Vec(0.6031752362132952, 0.7013438847413819, 0.06811502076305886)
 Vec(0.5927545912015951, 0.16203703104922063, 0.5261837731095815)
 Vec(0.2691780947158138, 0.2308430557128066, 0.41822600008827393)
 Vec(0.6981182380779529, 0.31059498134567376, 0.31290599081637627)
 ⋮
 Vec(0.2743833966333433, 0.2674326383631971, 0.4779061828120984)
 Vec(0.8084591461578993, 0.3783363872089769, 0.803287762206011)
 Vec(0.39434577244379077, 0.06107574915980607, 0.5381274230776456)
 Vec(0.45672289521705733, 0.5656510300306867, 0.9918422291735641)
 Vec(0.44262433268775714, 0.49775252980083673, 0.1835799236586707)
 Vec(0.370083666013055, 0.030708814285836605, 0.6105377013809613)
 Vec(0.23725207398748982, 0.19714290678933455, 0.5374240188227535)
 Vec(0.902433730310835, 0.49077683702697417, 0.20037975136428499)
 Vec(0.05892123059372134, 0.5285131978049287, 0.8454997086848125)