Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.1686298159315377, 0.9472490896726179, 0.893491346797086)
 Vec(0.2669068361630986, 0.425479331759276, 0.9388591538650903)
 Vec(0.6387482264754956, 0.0023985633844434417, 0.03466579851387053)
 Vec(0.438274193304951, 0.02255584850482506, 0.9695173508620425)
 Vec(0.09676875044872657, 0.2663613934805815, 0.2076342919779347)
 Vec(0.15905186676848027, 0.08706820122907977, 0.7528986273698584)
 Vec(0.07939113100342132, 0.6840481897795124, 0.5742995644711476)
 Vec(0.9246888944710525, 0.6858511029993152, 0.5002091643561186)
 Vec(0.5591138831581537, 0.49016152705656824, 0.9505128689299496)
 Vec(0.5023368778568921, 0.4857272193865637, 0.7518286493469468)
 ⋮
 Vec(0.38218783820494007, 0.9091427079870473, 0.22064592082162549)
 Vec(0.9931595274219227, 0.63601629266886, 0.4007908164843619)
 Vec(0.7999735035058799, 0.7800450646291092, 0.8044519869996455)
 Vec(0.30692562049258654, 0.07800046981754383, 0.6084173115072387)
 Vec(0.9679484926599932, 0.651706192774457, 0.8459982694670932)
 Vec(0.2893896746428811, 0.2250765848422438, 0.4873586111828411)
 Vec(0.7491402851582157, 0.6799042819740166, 0.6355610023322488)
 Vec(0.9202020322962295, 0.23064195571783308, 0.5570890386058802)
 Vec(0.45882865788413474, 0.558712871214356, 0.818709293705882)