Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.5060400614079117, 0.19061852408230628, 0.4300114083259601)
 Vec(0.9305683160335386, 0.4961396255626741, 0.6936425417066355)
 Vec(0.013495942439942454, 0.9901579513750063, 0.7025565917585335)
 Vec(0.39118139484390013, 0.7334595678186613, 0.9855580893712875)
 Vec(0.08127396939850551, 0.5993656344353617, 0.3198669733503827)
 Vec(0.32768018257765785, 0.6252774715282913, 0.568854116853633)
 Vec(0.6757382569181433, 0.7150369749727711, 0.17366199162499907)
 Vec(0.7750091076397272, 0.44952196237257613, 0.9397705356810552)
 Vec(0.15803037589915536, 0.19673491975727253, 0.5052775144763004)
 Vec(0.6070097757831897, 0.6681148258939841, 0.9653913343769677)
 ⋮
 Vec(0.4811759419101873, 0.27388804701754477, 0.600394449507455)
 Vec(0.5387912908855337, 0.8165056454386111, 0.45922536652712764)
 Vec(0.39879807837024095, 0.7590327698192408, 0.35944006019725505)
 Vec(0.7667645338323633, 0.7406644142041348, 0.029208710619274658)
 Vec(0.7607744063647541, 0.7608450923126461, 0.1769821596221215)
 Vec(0.8705704841183692, 0.44510134118181566, 0.3271655385082298)
 Vec(0.4773469401458733, 0.8198858276957953, 0.47082107340357393)
 Vec(0.7396236734455932, 0.07941162250130207, 0.7140440553101856)
 Vec(0.8005719307646457, 0.6601472386468465, 0.7172321972780218)