Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7132643916404031, 0.8490204566382242, 0.4530095296458291)
 Vec(0.5411066310768677, 0.19563016444936476, 0.4281115695580453)
 Vec(0.41643584527314437, 0.25236846429420423, 0.010585237247168422)
 Vec(0.49686228489734674, 0.48816378732204513, 0.7257409681100947)
 Vec(0.32235769235097544, 0.8624104564120294, 0.04094338276713072)
 Vec(0.5420103170403799, 0.4641038974440417, 0.47422780184223656)
 Vec(0.6353103727131422, 0.5247980490937948, 0.3986704761568627)
 Vec(0.7347609539282357, 0.34164809792859485, 0.03946299702297651)
 Vec(0.1152089896868852, 0.7928513988658206, 0.997629785588081)
 Vec(0.9696189147552272, 0.3772346763855221, 0.8885062980103692)
 ⋮
 Vec(0.06623565321359604, 0.2944923385447501, 0.1257049796804015)
 Vec(0.612737436174433, 0.7732484920725072, 0.05855161068282477)
 Vec(0.1303638028804871, 0.3033131183230261, 0.5838208353068902)
 Vec(0.5884764936238501, 0.7030666988624132, 0.6126073061970497)
 Vec(0.9803023117538706, 0.437002568773605, 0.8665096359151847)
 Vec(0.7137409869690865, 0.005550870883875847, 0.13624093952197014)
 Vec(0.1758058045675872, 0.42813456972728525, 0.6286078077779392)
 Vec(0.839720230960974, 0.9335529132752696, 0.9820749569203255)
 Vec(0.4573412079856938, 0.9949170559404931, 0.8308081333962332)