Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.5977401235596213, 0.42998942316266775, 0.14025220573708175)
 Vec(0.39678550734506524, 0.222091851161913, 0.294071928502065)
 Vec(0.8131059248272834, 0.44967551117178195, 0.7388704694683305)
 Vec(0.28891270583066786, 0.5628207843645414, 0.005467166886579222)
 Vec(0.20315022212657907, 0.9739731381261635, 0.774449750293193)
 Vec(0.4228941799447995, 0.4657779508032749, 0.13232291562820442)
 Vec(0.1642657160473655, 0.27122510663619626, 0.9887317830845385)
 Vec(0.023698653701991534, 0.7502792458588062, 0.1182224050912224)
 Vec(0.08657929368628425, 0.7290382576353998, 0.5024995926769946)
 Vec(0.008978834351873854, 0.5401559158151378, 0.5662474569054693)
 ⋮
 Vec(0.08962655684630927, 0.8816801571863403, 0.8316186501409)
 Vec(0.744874845949409, 0.5919760376404906, 0.0004165497550937536)
 Vec(0.9736959735679195, 0.18607909288656221, 0.9680464820289876)
 Vec(0.0472535236763415, 0.7577280838540332, 0.7627313355847549)
 Vec(0.4904424072810444, 0.764757392598345, 0.07245467508703318)
 Vec(0.25567969928316847, 0.7383697068741994, 0.31560186400538326)
 Vec(0.8096295735366285, 0.03880624808011368, 0.48105820992693005)
 Vec(0.975412577395407, 0.22657568610337464, 0.6440030580331884)
 Vec(0.5607857148707687, 0.4836748304783436, 0.8172560320574773)