Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.6680279024752079, 0.4712779062042588, 0.18399416565362203)
 Vec(0.2976496070549727, 0.7852251803402713, 0.4031630084998087)
 Vec(0.2631870409310344, 0.66710856070515, 0.1358771850458117)
 Vec(0.8013712108373986, 0.9514299709542006, 0.92818721706755)
 Vec(0.8936514084306606, 0.2589572149014022, 0.7108813810191857)
 Vec(0.8598129653239543, 0.7134845831033778, 0.8194760493212041)
 Vec(0.3658870711597495, 0.5138984323848833, 0.027564709971529378)
 Vec(0.024645658606630483, 0.4385700669351217, 0.9445410474435136)
 Vec(0.26528150673931894, 0.10684234273185178, 0.6244537514453041)
 Vec(0.1423747107524972, 0.06464844962761573, 0.8773639110854594)
 ⋮
 Vec(0.8054269016643308, 0.4583377601603339, 0.38924508970609994)
 Vec(0.37806638727394204, 0.4903747941779458, 0.4864070675439566)
 Vec(0.3930001197792289, 0.791684749168889, 0.10623010545701106)
 Vec(0.11002655215866364, 0.6311081760392876, 0.09071659240323338)
 Vec(0.2590792372223816, 0.6754693394670959, 0.31055622969567875)
 Vec(0.14577446323057486, 0.09719701564685612, 0.6622674218737398)
 Vec(0.2540005042790643, 0.7471485760155323, 0.2744704793378586)
 Vec(0.6692935246959834, 0.47916984582931, 0.7096683835436123)
 Vec(0.2444837015931257, 0.06929580131041013, 0.6898397914317254)