Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.2907642676108714, 0.4449333504222792, 0.1402918043055289)
 Vec(0.28835689277846865, 0.09119623023633072, 0.26733213101555464)
 Vec(0.9440985493153075, 0.6130668762019368, 0.07885297787662116)
 Vec(0.45499566523938006, 0.9881778742224743, 0.3940676925618052)
 Vec(0.551732761970373, 0.805132873138765, 0.49000375585065226)
 Vec(0.6801275306164416, 0.9702442924513668, 0.02758188564129216)
 Vec(0.8171626797540295, 0.15169170052787206, 0.8599733123975873)
 Vec(0.9070955034668924, 0.8913886975118577, 0.6047675428892586)
 Vec(0.554789930525479, 0.27679103952785544, 0.609752255139262)
 Vec(0.284592315999331, 0.4584449290870426, 0.0666566800019689)
 ⋮
 Vec(0.6046191065158926, 0.9880050210911449, 0.48212405898170685)
 Vec(0.20419605274687969, 0.056026336706311075, 0.3824960861726189)
 Vec(0.5114497626619158, 0.24933386587671214, 0.1174837599483386)
 Vec(0.02908950778194097, 0.7987246872155046, 0.5167275289992482)
 Vec(0.9408172465643194, 0.4569800608778931, 0.863661221569724)
 Vec(0.27423875730233926, 0.5478397293872349, 0.49311851559174946)
 Vec(0.7751925562718598, 0.9783516559082245, 0.6681979780293635)
 Vec(0.10102784144974919, 0.6625248605867651, 0.46512181091962734)
 Vec(0.09798464486598957, 0.3096530805641238, 0.8918647920412972)