Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7162870065770189, 0.6535079490313173, 0.015466794573737674)
 Vec(0.05003986524558868, 0.3875442056050935, 0.46513735087180996)
 Vec(0.8403866286735076, 0.8629678351557798, 0.9278702936956786)
 Vec(0.997928182472611, 0.9728860248076395, 0.4349998717533512)
 Vec(0.9380917954337636, 0.516364043109194, 0.045483353372362534)
 Vec(0.3916263725654221, 0.31672135433151427, 0.5867637027659401)
 Vec(0.7884509989489982, 0.3561532401180486, 0.49391140409858547)
 Vec(0.5300498642390672, 0.2956856250701808, 0.2721824970962752)
 Vec(0.9476863980189577, 0.5286696194684846, 0.3128922800195503)
 Vec(0.4859914680746432, 0.7864463128458032, 0.2721547378775995)
 ⋮
 Vec(0.826561987119035, 0.7696638178490103, 0.5666027375515298)
 Vec(0.3101408215540017, 0.14806933517362253, 0.6904340172232069)
 Vec(0.11720187486857647, 0.28914144507299544, 0.06036759261360114)
 Vec(0.1637979804724199, 0.950793591599577, 0.7559710171599777)
 Vec(0.06617055425576823, 0.8093770330586252, 0.8995885314053343)
 Vec(0.9292009719435764, 0.6935433497820912, 0.5441200052253239)
 Vec(0.10516021203472303, 0.9396586632664282, 0.7359950046697091)
 Vec(0.38004422349994893, 0.18244875090212143, 0.16422432909215512)
 Vec(0.8728594289300733, 0.8701428229788835, 0.36913658797456883)