Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.33136156828163144, 0.888090041799816, 0.7333062423701111)
 Vec(0.6584389171991051, 0.6448685905331482, 0.36404961736966324)
 Vec(0.5473341980775033, 0.20620459583358242, 0.7620778899433729)
 Vec(0.4810666448044536, 0.9625835682352165, 0.9168079305809508)
 Vec(0.5855486061217673, 0.4713289488571216, 0.8502107950413051)
 Vec(0.6451714094416651, 0.09537826606810296, 0.7997645252307389)
 Vec(0.031021599174277492, 0.35304047782781256, 0.3940714254381492)
 Vec(0.503944999669215, 0.8722692892848968, 0.7825075188380392)
 Vec(0.24816463870801064, 0.641211427248174, 0.9655434114740553)
 Vec(0.573421542338325, 0.33049639320246615, 0.0372385594221738)
 ⋮
 Vec(0.9827305106261358, 0.3288824384720278, 0.7886399492032745)
 Vec(0.6950574237837182, 0.32011772753673395, 0.638177326670407)
 Vec(0.22182038757282507, 0.2270918686080302, 0.9576362653368714)
 Vec(0.13612395096029029, 0.18145359680224726, 0.8880339230145733)
 Vec(0.12497784154886471, 0.5727626394015733, 0.42659229938361065)
 Vec(0.10677655885337267, 0.7113192517379522, 0.14495619620908573)
 Vec(0.7821767689820275, 0.9581729303046816, 0.12170894457236436)
 Vec(0.9958623199861925, 0.2331555111929895, 0.8410994195723701)
 Vec(0.8440708094307834, 0.5199499481071546, 0.3654943462502389)