Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.9445793954133367, 0.02020492331466106, 0.36754475820849764)
 Vec(0.5051017213336594, 0.026013629865669685, 0.19031736737962535)
 Vec(0.914942591137027, 0.03017724832464308, 0.6802254286941216)
 Vec(0.9179723942500859, 0.2884549443553487, 0.4553622454962031)
 Vec(0.8483891028933386, 0.4065036428081348, 0.6081919155012654)
 Vec(0.7719559847935467, 0.49750018508206584, 0.5757432728799655)
 Vec(0.4804807831212061, 0.8625745808610257, 0.34628936744620764)
 Vec(0.9995042136655385, 0.007895372588585636, 0.0008882432102322868)
 Vec(0.5667944567667311, 0.7819664692177494, 0.5798999959818754)
 Vec(0.02447484211538098, 0.06664676272818948, 0.1843314027225451)
 ⋮
 Vec(0.8658444586683848, 0.08375543218548909, 0.2045696322659012)
 Vec(0.21050873181150698, 0.3129888416103416, 0.7202469762798898)
 Vec(0.8956748535691647, 0.19762952284099544, 0.012137019389859338)
 Vec(0.17794372829604888, 0.32645806087147355, 0.1104778017449598)
 Vec(0.8866345372147187, 0.8396129835307251, 0.9792008468277775)
 Vec(0.6852594662280528, 0.6739285939722393, 0.7120017890142537)
 Vec(0.21418123870693917, 0.19695286238028342, 0.9118254456353589)
 Vec(0.5788223140267457, 0.25165982293752476, 0.2506402271955823)
 Vec(0.9862159827400333, 0.746540707057123, 0.25688986130268576)