Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.4187741521789895, 0.5725307746797628, 0.3126825678891548)
 Vec(0.32805495480486657, 0.08518163903725062, 0.32551737516878554)
 Vec(0.5911322241540654, 0.5263573987050038, 0.13442693818359086)
 Vec(0.03453994953367656, 0.018072225154671107, 0.8784074424773961)
 Vec(0.004294993479957432, 0.80938024659493, 0.8971137236131344)
 Vec(0.8797107897250412, 0.14177482910230732, 0.053282759288951764)
 Vec(0.363483372790648, 0.9704380091923448, 0.8567742485988569)
 Vec(0.5789338932518612, 0.8844709797308136, 0.04345261628069319)
 Vec(0.7308093649568703, 0.3753690788524683, 0.3542832357689578)
 Vec(0.18329465922730737, 0.16831570845543586, 0.4015043025196351)
 ⋮
 Vec(0.27135796610295426, 0.37384873065201296, 0.5956662992103946)
 Vec(0.46028779912564355, 0.010132567760099853, 0.6517165690103979)
 Vec(0.9160600284093112, 0.6873314636828318, 0.8635850928085194)
 Vec(0.9051944159093165, 0.05039773253779134, 0.889589272699066)
 Vec(0.04683506496265544, 0.22823658526640822, 0.33173233194810936)
 Vec(0.2532348805788184, 0.88097823930522, 0.009563504082338303)
 Vec(0.18450396725209628, 0.6730838222380923, 0.9834442532231983)
 Vec(0.27074288442117067, 0.5090392885628816, 0.2906951585386913)
 Vec(0.04203111735800047, 0.31747440487925116, 0.6452319706234703)