Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.6776572862557418, 0.3461229378347668, 0.8596125943577366)
 Vec(0.22202354221713894, 0.3579952239977573, 0.10516559111714563)
 Vec(0.23040353719671036, 0.7542788370950917, 0.6056896236130046)
 Vec(0.648126647722844, 0.3200626312899377, 0.7245324765935481)
 Vec(0.26257150701576926, 0.5027528243249334, 0.20078875085049874)
 Vec(0.28909012767541575, 0.6009631291074449, 0.010885818227519772)
 Vec(0.7178699962133807, 0.28409367935296204, 0.64500938629212)
 Vec(0.686844811256761, 0.8774679283242481, 0.027520684077111168)
 Vec(0.6147505362154652, 0.1560496221215255, 0.9016430461612787)
 Vec(0.38282927344182793, 0.03658197376502248, 0.5790141288300972)
 ⋮
 Vec(0.0674560291070262, 0.19654192767003031, 0.9084953198850918)
 Vec(0.8628379288377581, 0.15983158647623286, 0.4273567774446263)
 Vec(0.8810745120505582, 0.3817067740384449, 0.21227262954094794)
 Vec(0.5976421187046619, 0.5105463872245747, 0.19888874593312522)
 Vec(0.7343863662567424, 0.8384818512078985, 0.30285132268326154)
 Vec(0.5434237592097203, 0.38417697564952247, 0.7323037894643791)
 Vec(0.6820599451493911, 0.12099156987547688, 0.871123754071972)
 Vec(0.1586644196235507, 0.6114798596898747, 0.5043803131429481)
 Vec(0.014635163929858752, 0.4263053192859645, 0.6495760846876673)