Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.13610663014689206, 0.9085967551285468, 0.3120788053882122)
 Vec(0.2898140700087217, 0.13374118023823378, 0.8455712345216635)
 Vec(0.40680707434951113, 0.9232887483640274, 0.2204211876347536)
 Vec(0.03234807987921384, 0.8381077053443092, 0.7591949783771537)
 Vec(0.5369260084968542, 0.15246556707764325, 0.9419867104567562)
 Vec(0.5976207194565397, 0.42225437970111956, 0.9239692230846649)
 Vec(0.8581722869859839, 0.24228656469439303, 0.1365884712285621)
 Vec(0.741702001781971, 0.39554820903586807, 0.05201865377513604)
 Vec(0.4338981928636425, 0.8251933879324774, 0.4935144546033866)
 Vec(0.4950843502091433, 0.3770884417209075, 0.7376653770711301)
 ⋮
 Vec(0.34315768711843864, 0.9690768048053021, 0.1274156276707461)
 Vec(0.6951875819077118, 0.0323064360478571, 0.8285992063894014)
 Vec(0.7102183475872378, 0.7775932024919804, 0.9918924739685696)
 Vec(0.9778656595862104, 0.5384336741114698, 0.21312752361468912)
 Vec(0.1974897079261736, 0.3695443907643313, 0.6687149259249974)
 Vec(0.4614802218101459, 0.2765971902883543, 0.8314743364676617)
 Vec(0.3031732682159183, 0.3516591089987118, 0.28525781759124436)
 Vec(0.8944939182032203, 0.7132868815821656, 0.34898551832825053)
 Vec(0.7013189425646709, 0.05349883512301179, 0.12663591588596979)