Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.28750061571283936, 0.8736910475242192, 0.2752943971355718)
 Vec(0.4622851044599908, 0.2332254545873651, 0.8223339075797337)
 Vec(0.31733492739082614, 0.8423997595002943, 0.670240208260518)
 Vec(0.433724191251716, 0.6130094041596925, 0.8582580707733614)
 Vec(0.9048036979691938, 0.8470726296514163, 0.038313266121818956)
 Vec(0.7953532337484231, 0.32762007268381743, 0.8223869065457339)
 Vec(0.993836043504904, 0.7618365859665244, 0.7055322100534428)
 Vec(0.7066701901394993, 0.46489264492723137, 0.12437394628386655)
 Vec(0.9655726866080625, 0.21763950579863245, 0.6976497563845772)
 Vec(0.7103756771545139, 0.4060784344984518, 0.4728719673427466)
 ⋮
 Vec(0.45009885031297514, 0.1729539847117174, 0.7344913231259724)
 Vec(0.22983892103554537, 0.43823799269467356, 0.06337234263050773)
 Vec(0.8429824115301494, 0.44516568814824964, 0.9625915434474986)
 Vec(0.5687834038116163, 0.8055252059107116, 0.9700496355972132)
 Vec(0.22739596715173838, 0.8862411601269307, 0.4685178658038778)
 Vec(0.1705775704323207, 0.8258250162248212, 0.09206230283865369)
 Vec(0.5714799224550193, 0.6885736959964214, 0.30713230320310425)
 Vec(0.8770480494077918, 0.7443865516289945, 0.8525186072975592)
 Vec(0.8789262353819924, 0.5130707210007425, 0.8865357022413559)