Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.5174523363391501, 0.21794459439046088, 0.7490521504224298)
 Vec(0.3863530381510918, 0.7179982543971659, 0.48701674676249196)
 Vec(0.9851900727652159, 0.7920952298626398, 0.7140744997879264)
 Vec(0.43129846683831274, 0.024227994419785404, 0.22671758808849185)
 Vec(0.7788730055878235, 0.2576251251294217, 0.4909449706105917)
 Vec(0.3154463012444979, 0.9231977236213382, 0.48840774755273997)
 Vec(0.6302497079421299, 0.7451410712780488, 0.20157164525237525)
 Vec(0.2818078777044005, 0.7862482513838749, 0.9672803194864602)
 Vec(0.8559803108367665, 0.46442066943379146, 0.3686302387195902)
 Vec(0.9474017243155298, 0.5375707789424722, 0.10776777523257508)
 ⋮
 Vec(0.730285170727451, 0.7163804340170072, 0.7777981804051396)
 Vec(0.778953506018494, 0.12673346479554004, 0.7650302297594663)
 Vec(0.24948411798089654, 0.9969133549320129, 0.4376561986468839)
 Vec(0.01986390163579632, 0.2570147075279309, 0.3292212770331565)
 Vec(0.2690137330222976, 0.6884408338403805, 0.3974693893908795)
 Vec(0.2532091515028613, 0.6966253666900399, 0.7854757163420155)
 Vec(0.8343544859528965, 0.9468150398121352, 0.05095897016152784)
 Vec(0.07945278265027134, 0.9859456906328934, 0.0397475116695456)
 Vec(0.4767501387706031, 0.874237899624004, 0.5623745046475952)