Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.6620349016793622, 0.9072315960982192, 0.7636571546302399)
 Vec(0.8596236693883464, 0.5478890946731477, 0.6622341108115636)
 Vec(0.7846250558659811, 0.0949846378926743, 0.0034913088918535617)
 Vec(0.14180211776327378, 0.21431460723144047, 0.08155957190064167)
 Vec(0.04201524532823875, 0.4000088212704911, 0.1880601475537701)
 Vec(0.8209957085819363, 0.5551714886465493, 0.19878423628144326)
 Vec(0.21533475557893333, 0.417510250789508, 0.4264972742072797)
 Vec(0.8218172690302369, 0.2690967740063679, 0.7445005343416083)
 Vec(0.17475606806453803, 0.2927182754924713, 0.9630657070390207)
 Vec(0.9529788590868763, 0.19379425095288383, 0.07179664343226777)
 ⋮
 Vec(0.19471643648391523, 0.9147784013360002, 0.6711048129167124)
 Vec(0.8140158013070883, 0.8334109048572691, 0.751329069099346)
 Vec(0.9985595046691045, 0.2074868108870599, 0.021405977933150133)
 Vec(0.7877644420630829, 0.08966821824453519, 0.2751017153799955)
 Vec(0.9132725984092052, 0.5783744323854817, 0.2440949987658434)
 Vec(0.40332243504727583, 0.669825922982124, 0.020541965544664476)
 Vec(0.44039401192684147, 0.9173957190643242, 0.7904245270768335)
 Vec(0.972357389726746, 0.7067996345950696, 0.4832266153589786)
 Vec(0.7834106173220708, 0.9801082045987592, 0.09856597114491261)