Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.37043491312325416, 0.8789455193516171, 0.9249232825689432)
 Vec(0.9170562084578185, 0.19824271290498774, 0.43403380805774805)
 Vec(0.24690500964838413, 0.5949535477829744, 0.0648806133450297)
 Vec(0.6753011651022915, 0.5293641020268289, 0.05015120618175195)
 Vec(0.20370687543538424, 0.34908070213966913, 0.9257492458472288)
 Vec(0.8974654610788854, 0.2032492993547329, 0.658421110998389)
 Vec(0.43840301385033875, 0.6631018616760384, 0.008447271822322078)
 Vec(0.5202370400619774, 0.4930345297323637, 0.1158648113135583)
 Vec(0.9854484006479002, 0.23808080126029374, 0.28448907276192525)
 Vec(0.9975083445998585, 0.7577718555728664, 0.8154647677444679)
 ⋮
 Vec(0.9653917248605669, 0.32820435932586767, 0.8052439665959636)
 Vec(0.10118389218452073, 0.7170501998750733, 0.39685692503757286)
 Vec(0.8047109896215647, 0.4016493358056973, 0.5143071896424728)
 Vec(0.3500471272625424, 0.18357482938924552, 0.7551819222043681)
 Vec(0.47274062060972866, 0.27443407062126934, 0.1001111142210166)
 Vec(0.5296837123803309, 0.8418427745365034, 0.8046074026172504)
 Vec(0.03175447060534187, 0.7130050092888812, 0.2648348142933157)
 Vec(0.09484688988236512, 0.38912355609807836, 0.7007068406293822)
 Vec(0.006018539036388049, 0.7968711205281959, 0.6282358643545614)