Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7597778744854844, 0.03831459553710315, 0.6416711476283055)
 Vec(0.9113862569036008, 0.2664844147566724, 0.6028631339517463)
 Vec(0.3511182209324136, 0.5056476520021967, 0.27137189985206445)
 Vec(0.8509775668012731, 0.06043588686887047, 0.3360482823220351)
 Vec(0.6315599526598801, 0.35101615927762775, 0.8657511692790728)
 Vec(0.4932717332473231, 0.5566611451116447, 0.08249645376327275)
 Vec(0.3336245927857151, 0.05809956625329915, 0.7960714058919984)
 Vec(0.5973147819441298, 0.5192598664099815, 0.31692802869347714)
 Vec(0.11615654479974824, 0.2859256153304389, 0.048225577804656417)
 Vec(0.45403394543807374, 0.35881765427137124, 0.6765842572125271)
 ⋮
 Vec(0.530967108523823, 0.9909839405795415, 0.8690645978440459)
 Vec(0.9703148625794744, 0.6922695711738908, 0.34183177297764533)
 Vec(0.9006234031650661, 0.746660810879239, 0.053371697124665385)
 Vec(0.518338960504526, 0.6809220962715384, 0.060125092601514774)
 Vec(0.3064819622722008, 0.14125150325468705, 0.12881054304372042)
 Vec(0.031169420812750315, 0.10781189034926952, 0.031611485777400916)
 Vec(0.14167904802751796, 0.7411091617021653, 0.4979948982878999)
 Vec(0.7812916879142631, 0.06477682382701189, 0.47157909271607656)
 Vec(0.51663226831719, 0.6866542031209623, 0.13212385311857044)