Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.3845876166738108, 0.3150264366690345, 0.5245845325639137)
 Vec(0.15725100120734292, 0.529782281415556, 0.49814899869309737)
 Vec(0.4306622463814227, 0.8233003285379691, 0.23150995692629328)
 Vec(0.22297362077343819, 0.18512231900835419, 0.9658781880890374)
 Vec(0.06213200516374473, 0.9542520297669427, 0.9136157881677279)
 Vec(0.4404511302627594, 0.8798715684056524, 0.7774795631046545)
 Vec(0.09338255249885374, 0.6159193509066405, 0.23973576037383493)
 Vec(0.23766109641369604, 0.8983871164524516, 0.5792922561196133)
 Vec(0.4317675367750283, 0.33286483525358324, 0.48517016672651225)
 Vec(0.3225071281614623, 0.7559463706096934, 0.46161020473487013)
 ⋮
 Vec(0.4237884161920168, 0.41221732982233106, 0.019980399753933842)
 Vec(0.11545369099715685, 0.8753428840451445, 0.5522017630514874)
 Vec(0.6042841149379822, 0.3698295361662589, 0.18036817339128008)
 Vec(0.8363742508403618, 0.7178065860348938, 0.7994486223448751)
 Vec(0.947082971271642, 0.32852918042225165, 0.6282823640440724)
 Vec(0.08809403945376248, 0.18416063424032458, 0.02178007427049289)
 Vec(0.26960376333383074, 0.6049405340705173, 0.988157759223403)
 Vec(0.17043907083228094, 0.5623738576130556, 0.8993593508036627)
 Vec(0.948082693979532, 0.06357027609234978, 0.6818117451003465)