Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.8979701823062503, 0.7644680886268722, 0.42153979803092634)
 Vec(0.19658134451467346, 0.5196098109674088, 0.6947622854182925)
 Vec(0.7599843953580699, 0.099200176254298, 0.09498966828372069)
 Vec(0.5169935676258968, 0.028711316439195556, 0.4155920004877609)
 Vec(0.3690717409935219, 0.6697747367656085, 0.1653796714794874)
 Vec(0.7734123402722651, 0.6663608068451468, 0.06730030254137442)
 Vec(0.7557572628527749, 0.9760889914231963, 0.46011930414031776)
 Vec(0.1540770974029827, 0.062290778255653434, 0.9015094525095678)
 Vec(0.9389045298858232, 0.49717200613825807, 0.5010509386878863)
 Vec(0.6480639799081162, 0.1649294941913223, 0.20198130902930345)
 ⋮
 Vec(0.6692176548146211, 0.8704211512864152, 0.9742536095716837)
 Vec(0.22618237297594268, 0.00939833848930094, 0.9114393981995249)
 Vec(0.19015246125525398, 0.6209080135342708, 0.12785213901625547)
 Vec(0.5225161521400687, 0.5505253580540773, 0.05686721154920582)
 Vec(0.21780397214174974, 0.049665309044486405, 0.23582913905619418)
 Vec(0.7305944580702219, 0.6878358273239282, 0.6633843151784076)
 Vec(0.17554764439082382, 0.17726280353010293, 0.781494671898317)
 Vec(0.4363849057904904, 0.45204188403111745, 0.8575145197075569)
 Vec(0.7911041578935163, 0.6951924711186312, 0.20458181250725505)