Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.7533794574538278, 0.5584030793814745, 0.3794458868855787)
 Vec(0.5365491078726947, 0.08557156494386553, 0.9405634443177228)
 Vec(0.415870159856616, 0.04530844936549805, 0.37361754590827856)
 Vec(0.09768330595611696, 0.49723646253643894, 0.7557728737708024)
 Vec(0.5128540379856541, 0.932330491269255, 0.07325007927978766)
 Vec(0.6205238369671047, 0.791013187922746, 0.5764360479541752)
 Vec(0.758778472010898, 0.45251576135867455, 0.9465069534134877)
 Vec(0.5059903029377323, 0.8285075056030526, 0.11920541448979649)
 Vec(0.8600918023625638, 0.7390548774893451, 0.007704787796493084)
 Vec(0.8037666420172378, 0.10791995153462752, 0.9646959648179427)
 ⋮
 Vec(0.742163668910366, 0.7619217284097468, 0.7277497045668513)
 Vec(0.591569910513564, 0.21047641636517245, 0.3151944204583539)
 Vec(0.7058011252978547, 0.39264522596721196, 0.35148094162616406)
 Vec(0.9565743565255377, 0.8847740878120376, 0.8585836484355162)
 Vec(0.3974146315700694, 0.9845156492393778, 0.2907587434435759)
 Vec(0.05019082295311572, 0.5416328878986121, 0.41245985035695754)
 Vec(0.45020662387401544, 0.08712037950913076, 0.45998236585841645)
 Vec(0.41202562219468264, 0.7567948189975365, 0.8828128927692287)
 Vec(0.5757161930511698, 0.195324960022923, 0.3195330500452055)