Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.42643030618489464, 0.75075356302831, 0.18708064359952759)
 Vec(0.9931552720544744, 0.3893986390466102, 0.533847372743995)
 Vec(0.4811923218332118, 0.24215834968146432, 0.41414851722111923)
 Vec(0.7273644308513562, 0.8253524990963856, 0.46825095076465806)
 Vec(0.7156644706790626, 0.7187246344322903, 0.9580677875803725)
 Vec(0.4970534103178148, 0.7402795222152053, 0.7750906773782155)
 Vec(0.8761497119239198, 0.56567768355402, 0.8406740905585381)
 Vec(0.13391873134954313, 0.4426786085351868, 0.884847187663038)
 Vec(0.24327943845739497, 0.7081364584163212, 0.32539832209493214)
 Vec(0.8510766206112199, 0.41807682924204803, 0.3044494698783281)
 ⋮
 Vec(0.09012963980256206, 0.14096708160560878, 0.8106454330786816)
 Vec(0.9806213021168249, 0.9807818430812557, 0.7205142638202309)
 Vec(0.40676905116815365, 0.07231628097596765, 0.8337372646979427)
 Vec(0.11859233644010558, 0.8755630231940533, 0.5660445662100057)
 Vec(0.1105414743108376, 0.599856020662256, 0.09170942290638251)
 Vec(0.05403835757838116, 0.13418896734028907, 0.016926634864036116)
 Vec(0.7034357014699824, 0.5664587106439112, 0.3794825857638646)
 Vec(0.9670842219442639, 0.012324119133841904, 0.17057957429383208)
 Vec(0.9130451701603691, 0.22261030360288692, 0.7271677060229582)