Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.3819861362755813, 0.3711758023610173, 0.022985383459646402)
 Vec(0.5499049490375129, 0.16939342398009583, 0.48373681682694447)
 Vec(0.7270901312320824, 0.6157658190768387, 0.5346403070193316)
 Vec(0.7341683522010652, 0.7304268072353927, 0.29105988776601843)
 Vec(0.5138662066977236, 0.1766371841640083, 0.4827339608861876)
 Vec(0.056687529911434775, 0.011580973089704072, 0.1767749595785989)
 Vec(0.7646719223350591, 0.29359422008474356, 0.7179573421222688)
 Vec(0.3690647403608034, 0.2201076846320681, 0.9276616035221498)
 Vec(0.1430707627770743, 0.019737793880014398, 0.6791186342699133)
 Vec(0.3302407052086227, 0.18267208289860204, 0.1531122704544403)
 ⋮
 Vec(0.7100303440646578, 0.4359661722492286, 0.17256115764878277)
 Vec(0.2933257966897709, 0.32831069417790526, 0.6476225267297454)
 Vec(0.3248191171268372, 0.16784628959511994, 0.12965825740302417)
 Vec(0.46458727889241214, 0.22066116256142299, 0.48823574745402265)
 Vec(0.2269017827779174, 0.7669299096121117, 0.8095385313858692)
 Vec(0.5645727745143891, 0.6290495528308369, 0.6051537803960941)
 Vec(0.9060416663085196, 0.0371094139885898, 0.8138243201364647)
 Vec(0.8079010856811587, 0.889412217975845, 0.6570161934345432)
 Vec(0.6699332932585402, 0.2706703784373393, 0.9373064541945063)