Vectors

Meshes.Vec — Type

Vec(x₁, x₂, ..., xₙ)
Vec((x₁, x₂, ..., xₙ))
Vec{Dim,T}(x₁, x₂, ..., xₙ)
Vec{Dim,T}((x₁, x₂, ..., xₙ))

A geometric vector in Dim-dimensional space with coordinates of type T for linear algebra.

By default, integer coordinates are converted to float.

A vector can be obtained by subtracting two Point objects.

Examples

A = Point(0.0, 0.0)
B = Point(1.0, 0.0)
v = B - A

# 2D vectors
Vec(0.0, 1.0) # double precision as expected
Vec(0f0, 1f0) # single precision as expected
Vec(0, 0) # integer is converted to float by design
Vec2(0, 1) # explicitly ask for double precision
Vec2f(0, 1) # explicitly ask for single precision

# 3D vectors
Vec(1.0, 2.0, 3.0) # double precision as expected
Vec(1f0, 2f0, 3f0) # single precision as expected
Vec(1, 2, 3) # integer is converted to float by design
Vec3(1, 2, 3) # explicitly ask for double precision
Vec3f(1, 2, 3) # explicitly ask for single precision

Notes

A Vec is a subtype of StaticVector from StaticArrays.jl

Type aliases are Vec1, Vec2, Vec3, Vec1f, Vec2f, Vec3f

rand(Vec3, 100)

100-element Vector{Vec3}:
 Vec(0.17345159069151594, 0.6303509590821486, 0.33195127377955214)
 Vec(0.4691033872445465, 0.10588502837905467, 0.2161609948835631)
 Vec(0.023642210792855556, 0.589011079684675, 0.3965136966026359)
 Vec(0.6572530520836329, 0.06598141180638395, 0.05697796669272859)
 Vec(0.0644992515552456, 0.383238020953375, 0.6840163904032666)
 Vec(0.879296680615322, 0.030005984759080118, 0.8899680980133587)
 Vec(0.4795642245663687, 0.7043989587480486, 0.5228301254399372)
 Vec(0.6709792343204128, 0.09337442618875957, 0.41956473807881023)
 Vec(0.6417884142634378, 0.7281697359269962, 0.39067424603825507)
 Vec(0.5081536438547957, 0.8175593004525273, 0.07971232396789107)
 ⋮
 Vec(0.5634121554781086, 0.6570382163189258, 0.30461456841780954)
 Vec(0.42009921104465486, 0.9740406672905958, 0.696686342096865)
 Vec(0.24506072592113737, 0.26832354892435484, 0.5910202985124292)
 Vec(0.4579575663453178, 0.4915913246454843, 0.12872325139122054)
 Vec(0.6967359575364164, 0.6570496237945881, 0.46078544390660503)
 Vec(0.5512359540218418, 0.38612682822753275, 0.16048953465673532)
 Vec(0.3395111467275689, 0.4665484793804193, 0.17060037460423694)
 Vec(0.3234274624664598, 0.6054987811465208, 0.4040072609141502)
 Vec(0.04467863700785757, 0.1175484730192059, 0.26619731387899737)